ft-#77: Remove unused dir
