--- src/services/transactionDetails/index.ts
@@ -1,16 +1,0 @@
-// import { axiosInstance } from "@utils/axiosInstance";
-// import { ETransactionAllType } from "@services/transaction/types";
-// import { TTransactionDetails } from "./types";
-// import { CLIENT_ROUTES } from "@constants/routes";
-
-// const getOneTransaction = async (type: ETransactionAllType, id: string) => {
-//   const response = await axiosInstance({
-//     method: "GET",
-//     url: `${CLIENT_ROUTES.allTransactions}/${type}/${id}`,
-//   });
-
-//   console.log("data", response.data.data);
-//   return response.data.data as TTransactionDetails;
-// };
-
-// export default getOneTransaction;
